Source -> Stream. Create the Store concept and move netCDF writing there
